Conditionally include google analytics
In order to prevent GA from tracking myself during development, I'm
conditionally adding the code if window.location.hostname ===
www.jakewiesler.com.

This may not be the "best" way to do things, but it should suffice for
now. I've also added a filter in GA to block my IP as well.